+++ src/utils/__tests__/data.spec.ts
@@ -1,9 +1,13 @@
 import { describe, it, expect } from 'vitest'
 
+import type {Format} from '@/types/app'
+
 import {
   getExtent,
   getExtents,
   getFormat,
+  getFormatByExtension,
+  getFormatByType,
   getFormats,
   getIdeas,
   getIndividuals,
@@ -29,11 +33,12 @@ const checkExtent = {
   }
 }
 
-const checkFormat = {
+const checkFormat: Format = {
   slug: 'gpkg',
   name: 'GeoPackage',
   version: '1.2',
   extensions: ['.gpkg'],
+  mediaTypes: ['application/geopackage+sqlite3'],
   url: 'https://www.iana.org/assignments/media-types/application/geopackage+sqlite3'
 }
 
@@ -64,6 +69,26 @@ describe('getExtents', () => {
 describe('getFormat', () => {
   it('loads expected format', () => {
     expect(getFormat(checkFormat.slug)).toEqual(checkFormat)
+  })
+})
+
+describe('getFormatByExtension', () => {
+  it('loads expected format', () => {
+    expect(getFormatByExtension('.gpkg')).toEqual(checkFormat)
+  })
+
+  it('does not load unexpected format', () => {
+    expect(getFormatByExtension('.foo')).toBeUndefined()
+  })
+})
+
+describe('getFormatByType', () => {
+  it('loads expected format', () => {
+    expect(getFormatByType('application/geopackage+sqlite3')).toEqual(checkFormat)
+  })
+
+  it('does not load unexpected format', () => {
+    expect(getFormatByType('application/foo')).toBeUndefined()
   })
 })