+++ src/utils/data.ts
@@ -28,6 +28,16 @@ export const getFormat = (slug: string): Format => {
   return (formatsData.formats as Record<string, Format>)[slug]
 }
 
+export const getFormatByExtension = (extension: string): Format | undefined => {
+  const match = Object.values(formatsData.formats).find((format: Format) => format.extensions?.includes(extension))
+  if (match) return match
+}
+
+export const getFormatByType = (type: string): Format | undefined => {
+  const match = Object.values(formatsData.formats).find((format: Format) => format.mediaTypes?.includes(type))
+  if (match) return match
+}
+
 export const getFormats = (): Format[] => {
   return Object.values(formatsData.formats)
 }