+++ CHANGELOG.md
@@ -41,6 +41,9 @@
 * Refactoring usage constraint construction to utils module from licence component
 * Minor refactoring for citation component to remove unnecessary null identifier logic
 * Refactoring service component to generate distribution option via util functions
+* Refactoring download component to generate distribution option via util functions
+* Refactoring download component to determine distribution format via media-type where available
+* Selecting relevant distributor organisation based on resource type
 * Refactoring formats to make version optional
 
 ## [0.5.3] - 2023-12-30